requirements.txt
torch~=1.6.0
numpy~=1.19.1
scikit-learn
pymethylprocess~=0.1.6
matplotlib~=3.3.1
pandas~=1.1.2
torchvision~=0.7.0
click~=6.7
networkx~=2.4
dask~=2.18.0
shap~=0.35.0
setuptools~=50.3.0
tensorflow~=2.3.0
tensorflow-gpu~=2.3.0
pillow~=7.2.0